requirements.txt
scipy
numpy
pytorch
torchvision
tqdm
pywavelets
